Split architectures in stable/experimental/deprecated (#79)
---------
