+++ .github/PULL_REQUEST_TEMPLATE/pull_request_new-architecture.md
@@ -1,0 +1,14 @@
+<!-- Describe your new architecture briefly here -->
+
+
+
+# New architecture TODOs
+
+- [ ] Add your architecture to the experimental folder
+  `src/metatensor/models/experimental/<architecture_name>`
+- [ ] Add default hyperparameter file to
+  `src/metatensor/models/cli/conf/architecture/experimental.<architecture_name>.yaml`
+- [ ] Add a `.yml` file into github workflows `.github/workflow/<architecture_name>.yml`
+- [ ] Architecture dependencies entry in the `optional-dependencies` section in the
+  `pyproject.toml`
+- [ ] Tests: torch-scriptability, basic functionality (invariance, fitting, prediction)
