--- .github/PULL_REQUEST_TEMPLATE/pull_request_new-model.md
@@ -1,10 +1,0 @@
-<!-- Describe your new model briefly here -->
-
-
-
-# New model TODOs
-
-- [ ] Add default hyperparameter file to `models/cli/conf/architecture/<YOUR_model>.yaml`
-- [ ] Tests: torch-scriptability, basic functionality (invariance, fitting, prediction)
-- [ ] Add a `.yml` file into github workflows `.github/workflow/<YOUR_model>.yml`
-- [ ] Entry in the `optional-dependencies` section in the `pyproject.toml`
