+++ MANIFEST.in
@@ -3,6 +3,7 @@ graft src
 include LICENSE
 include README.md
 
+prune developer
 prune docs
 prune examples
 prune tests
