--- src/metatensor/models/cli/conf/architecture/soap_bpnn.yaml
@@ -1,32 +1,0 @@
-# default hyperparameters for the SOAP-BPNN model
-name: soap_bpnn
-
-model:
-  soap:
-    cutoff: 5.0
-    max_radial: 8
-    max_angular: 6
-    atomic_gaussian_width: 0.3
-    radial_basis:
-      Gto: {}
-    center_atom_weight: 1.0
-    cutoff_function:
-      ShiftedCosine:
-        width: 1.0
-    radial_scaling:
-      Willatt2018:
-        rate: 1.0
-        scale: 2.0
-        exponent: 7.0
-
-  bpnn:
-    num_hidden_layers: 2
-    num_neurons_per_layer: 32
-    activation_function: SiLU
-
-training:
-  batch_size: 8
-  num_epochs: 100
-  learning_rate: 0.001
-  log_interval: 10
-  checkpoint_interval: 25
