+++ src/metatensor/models/experimental/soap_bpnn/tests/__init__.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+DATASET_PATH = str(
+    Path(__file__).parent.resolve()
+    / "../../../../../../tests/resources/qm9_reduced_100.xyz"
+)
