+++ src/metatensor/models/experimental/soap_bpnn/tests/test_torchscript.py
@@ -1,0 +1,43 @@
+import torch
+from metatensor.torch.atomistic import ModelCapabilities, ModelOutput
+
+from metatensor.models.experimental.soap_bpnn import DEFAULT_HYPERS, Model
+
+
+def test_torchscript():
+    """Tests that the model can be jitted."""
+
+    capabilities = ModelCapabilities(
+        length_unit="Angstrom",
+        species=[1, 6, 7, 8],
+        outputs={
+            "energy": ModelOutput(
+                quantity="energy",
+                unit="eV",
+            )
+        },
+    )
+    soap_bpnn = Model(capabilities, DEFAULT_HYPERS["model"]).to(torch.float64)
+    torch.jit.script(soap_bpnn, {"energy": soap_bpnn.capabilities.outputs["energy"]})
+
+
+def test_torchscript_save():
+    """Tests that the model can be jitted and saved."""
+
+    capabilities = ModelCapabilities(
+        length_unit="Angstrom",
+        species=[1, 6, 7, 8],
+        outputs={
+            "energy": ModelOutput(
+                quantity="energy",
+                unit="eV",
+            )
+        },
+    )
+    soap_bpnn = Model(capabilities, DEFAULT_HYPERS["model"]).to(torch.float64)
+    torch.jit.save(
+        torch.jit.script(
+            soap_bpnn, {"energy": soap_bpnn.capabilities.outputs["energy"]}
+        ),
+        "soap_bpnn.pt",
+    )
