--- src/metatensor/models/soap_bpnn/__init__.py
@@ -1,2 +1,0 @@
-from .model import Model, DEFAULT_HYPERS  # noqa: F401
-from .train import train  # noqa: F401
