--- src/metatensor/models/soap_bpnn/tests/__init__.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-DATASET_PATH = str(
-    Path(__file__).parent.resolve()
-    / "../../../../../tests/resources/qm9_reduced_100.xyz"
-)
