+++ tests/resources/options.yaml
@@ -1,5 +1,5 @@
 architecture:
-  name: soap_bpnn
+  name: experimental.soap_bpnn
   training:
     batch_size: 2
     num_epochs: 1
