+++ tests/utils/test_output_gradient.py
@@ -6,7 +6,7 @@ import rascaline.torch
 import torch
 from metatensor.torch.atomistic import ModelCapabilities, ModelOutput
 
-from metatensor.models import soap_bpnn
+from metatensor.models.experimental import soap_bpnn
 from metatensor.models.utils.data import read_structures
 from metatensor.models.utils.output_gradient import compute_gradient
