+++ tox.ini
@@ -79,7 +79,7 @@ passenv = *
 deps =
     pytest
 commands =
-    pytest --import-mode=append {posargs} src/metatensor/models/soap_bpnn/tests/
+    pytest --import-mode=append {posargs} src/metatensor/models/experimental/soap_bpnn/tests/
 
 [testenv:docs]
 description = builds the documentation with sphinx
